Actualización de observaciones

+++ PHP/verclientesadmin.php
@@ -4,9 +4,9 @@ session_start();
 
 // Establecer la conexión a la base de datos
 $servername = "localhost";
-$dbUsername = "root";
-$dBPassword = "";
-$dbName = "PlanchaduriaLomaLinda";
+$dbUsername = "id20635539_planchadurialomalinda";
+$dBPassword = "Plaloli123!";
+$dbName = "id20635539_dbplanchaduria";
 
 $conn = mysqli_connect($servername,$dbUsername,$dBPassword,$dbName);
 
@@ -24,20 +24,26 @@ $fClientes = mysqli_query($conn, $cClientes);
 	<link rel="stylesheet" type="text/css" href="CSS/desginverclientes.css">
 </head>
 <body>
-	<header>
-		<h1>Administración de clientes</h1>
-		<nav>
-			<ul>
-				<li><a href="sitioweb/administrador/seccion/productos.php">Productos</a></li>
-				<li><a href="pedidosadmin.php">Pedidos</a></li>
-				<a class="nav-item nav-link" href="sitioweb/administrador/inicio.php">Volver</a>
-				<li><a href="header.php">Cerrar sesión</a></li>
-			</ul>
-		</nav>
-	</header>
+	<head>
+		<title>Administración de clientes</title>
+		<meta charset="utf-8">
+		<link rel="stylesheet" type="text/css" href="CSS/desginverclientes.css">
+		<script>
+	    	function confirmarEliminacion() {
+	        	var respuesta = confirm("¿Está seguro de que desea eliminar al cliente?");
+	        	if (respuesta == true) {
+	            	return true;
+	        	} else {
+	            	return false;
+	        	}
+	    	}
+		</script>
+	</head>
+
 
 	<main>
 		<h2>Lista de clientes</h2>
+		<a href="sitioweb/administrador/inicio.php">Regresar</a>
 		<table>
 			<thead>
 				<tr>
@@ -54,8 +60,7 @@ $fClientes = mysqli_query($conn, $cClientes);
 					<td><?php echo $fila['uIdClientes']; ?></td>
 					<td><?php echo $fila['CorreoCliente']; ?></td>
 					<td>
-						<a href="editarperfiladmin.php?id=<?php echo $fila['Id_Cliente']; ?>">Editar</a>
-						<a href="eliminarclientesadmin.php?id=<?php echo $fila['Id_Cliente']; ?>">Eliminar</a>
+						<a href="eliminarclientesadmin.php?id=<?php echo $fila['Id_Cliente']; ?>" onclick="return confirmarEliminacion();">Eliminar</a>
 					</td>
 				</tr>
 				<?php } ?>
